+++ foxes/models/turbine_types/__init__.py
@@ -9,3 +9,4 @@ from .CpCt_from_two import CpCtFromTwo
 from .null_type import NullType
 from .wsrho2PCt_from_two import WsRho2PCtFromTwo
 from .wsti2PCt_from_two import WsTI2PCtFromTwo
+from .lookup import FromLookupTable